requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.1
requests>=2.31.0
beautifulsoup4>=4.12.3
selenium>=4.15.2
aiohttp>=3.9.3
requests
mcstatus
python_aternos
exaroton 
playwright
bs4
selenium
